requirements.txt
# Core dependencies
openai
python-dotenv>=1.0.0
pydantic>=2.4.0
typing-extensions>=4.7.0

# Testing dependencies
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-mock>=3.10.0
pytest-xdist>=3.3.1
responses>=0.23.1

# Development dependencies
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.3.0